+++ lib/main.dart
@@ -1,8 +1,6 @@
-import 'package:final_application/screens/AuthScreen/plashScreen.dart';
+import 'package:final_application/screens/AuthScreen/splash_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-
-
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -15,7 +13,7 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Home(),
-    ) ;
+    );
   }
 }
 
@@ -25,27 +23,23 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-
   final Future<FirebaseApp> _initialization = Firebase.initializeApp();
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-      future: _initialization,
-      builder: (context, snapshot) {
-        if(snapshot.hasError){
-          return Container();
-        }
-        
-        if(snapshot.connectionState == ConnectionState.done){
-          return Splash();
+        future: _initialization,
+        builder: (context, snapshot) {
+          if (snapshot.hasError) {
+            return Container();
+          }
 
-          
-        }
+          if (snapshot.connectionState == ConnectionState.done) {
+            return Splash();
+          }
 
-        return Center(
-          child: CircularProgressIndicator(),
-        );
-      });
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        });
   }
-  
 }
